refactor: refactor to minimize usage of bigdecimal and replace with f64

+++ tests/token_price_hash.rs
@@ -26,7 +26,7 @@ async fn test_setting_price_with_token_price_hash() -> Result<()> {
     // get address for WETH
     let weth_token = token_data.get("WETH").unwrap();
     let weth_price = weth_token.get_token_oracle_price(&client).await?;
-    let new_price = BigDecimal::from(2) * &weth_price;
+    let new_price = 2.0 * &weth_price;
 
     set_saved_token_price(weth_token.address.as_str(), new_price.clone()).await?;
